COMP: Fix -Wsign-compare warnings
This commit fixes the following warnings:

/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx:747:13: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     if(event==this->Internal->IOConnector->NewDeviceEvent)
             ^
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx: In member function ‘virtual void vtkMRMLIGTLConnectorNode::OnNodeReferenceAdded(vtkMRMLNode::vtkMRMLNodeReference*)’:
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx:1021:41: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
       for (int typeIndex = 0; typeIndex < deviceTypes.size(); typeIndex++)
                                         ^
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx: In member function ‘int vtkMRMLIGTLConnectorNode::RegisterOutgoingMRMLNode(vtkMRMLNode*, const char*)’:
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx:1282:39: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int typeIndex = 0; typeIndex < deviceTypes.size(); typeIndex++)
                                       ^
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx: In member function ‘void* vtkMRMLIGTLConnectorNode::CreateDeviceForOutgoingMRMLNode(vtkMRMLNode*)’:
/tmp/SlicerOpenIGTLink/OpenIGTLinkIF/MRML/vtkMRMLIGTLConnectorNode.cxx:1646:39: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     for (int typeIndex = 0; typeIndex < deviceTypes.size(); typeIndex++)
                                       ^

+++ OpenIGTLinkIF/Logic/vtkSlicerOpenIGTLinkIFLogic.cxx
@@ -102,7 +102,7 @@ vtkSlicerOpenIGTLinkIFLogic::vtkSlicerOpenIGTLinkIFLogic()
   this->RestrictDeviceName = 0;
   
   std::vector<std::string> deviceTypes = this->Internal->DeviceFactory->GetAvailableDeviceTypes();
-  for (int typeIndex = 0; typeIndex<deviceTypes.size();typeIndex++)
+  for (size_t typeIndex = 0; typeIndex<deviceTypes.size();typeIndex++)
   {
     this->Internal->MessageDeviceList.push_back(this->Internal->DeviceFactory->GetCreator(deviceTypes[typeIndex])->Create(""));
   }
